refrac: changed location of status styling and renamed
